integrate consurf and moda color scripts into one
