+++ sample-consurf-consurf-coloring-script.pml
