+++ sample_moda-moda-coloring-script.pml
